+++ src/test/java/ch/hsr/adit/application/controller/UserControllerIT.java
@@ -162,11 +162,7 @@ public class UserControllerIT {
     assertEquals(200, response.statusCode);
     assertNotNull(json.get("id"));
     assertNotNull(json2.get("id"));
-    assertEquals((double)json.get("id") + 1, json2.get("id"));
-
-
-    
-    
+    assertEquals((Long)json.get("id") + 1, json2.get("id"));
   }
 
 }
